rename the digitizer faux py files by removing the leading "faux"
