+++ bapsflib/_hdf/maps/digitizers/tests/test_lecroy.py
@@ -4,7 +4,7 @@ from unittest import mock
 
 from bapsflib._hdf.maps.digitizers.lecroy import HDFMapDigiLeCroy180E
 from bapsflib._hdf.maps.digitizers.tests.common import DigitizerTestCase
-from bapsflib.faux.digitizers.fauxlecroy180e import FauxLeCroy180E
+from bapsflib.faux.digitizers.lecroy180e import FauxLeCroy180E
 from bapsflib.utils.exceptions import HDFMappingError
 from bapsflib.utils.warnings import HDFMappingWarning
