+++ bapsflib/_hdf/utils/tests/test_helpers.py
@@ -18,7 +18,7 @@ from h5py import Group
 from numpy.lib import recfunctions as rfn
 
 from bapsflib._hdf.maps.controls.waveform import HDFMapControlWaveform
-from bapsflib.faux.digitizers.fauxlecroy180e import FauxLeCroy180E
+from bapsflib.faux.digitizers.lecroy180e import FauxLeCroy180E
 from bapsflib._hdf.utils.file import File
 from bapsflib._hdf.utils.helpers import (
     build_shotnum_dset_relation,
