+++ bapsflib/faux/digitizers/lecroy180e.py
