+++ bapsflib/faux/digitizers/sis3301.py
