+++ bapsflib/faux/digitizers/siscrate.py
